Improve type hinting, error handling, and code quality across multiple modules (#53)

+++ src/__init__.py
@@ -5,7 +5,7 @@
 import asyncio
 import os
 import shutil
-import sys
+from datetime import datetime
 
 from pytdbot import Client, types
 
@@ -14,15 +14,7 @@ from src.helpers import call, db, start_clients
 from src.modules.jobs import InactiveCallManager
 
 __version__ = "1.1.8"
-
-# Enforce a minimum Python version
-if sys.version_info < (3, 10):
-    raise RuntimeError(
-        "\n\n❌ Your Python version is too old!\n"
-        "   This project requires Python 3.10 or newer to run.\n"
-        "   Please upgrade Python:\n"
-        "     🔗 https://www.python.org/downloads/\n"
-    )
+StartTime = datetime.now()
 
 
 class Telegram(Client):
@@ -50,7 +42,8 @@ class Telegram(Client):
         await call.register_decorators()
         await self.call_manager.start_scheduler()
         await super().start()
-        self.logger.info("✅ Bot started successfully.")
+        self.logger.info(f"Bot started in {datetime.now() - StartTime} seconds.")
+        self.logger.info(f"Version: {__version__}")
 
     async def stop(self) -> None:
         shutdown_tasks = [
@@ -58,10 +51,12 @@ class Telegram(Client):
             self.call_manager.stop_scheduler(),
             super().stop(),
         ]
-        await asyncio.gather(*shutdown_tasks, return_exceptions=False)
+        await asyncio.gather(*shutdown_tasks)
 
     @staticmethod
     def _check_config() -> None:
+        if not config.API_ID or not config.API_HASH or not config.TOKEN:
+            raise ValueError("API_ID, API_HASH and TOKEN are required")
         if config.IGNORE_BACKGROUND_UPDATES and os.path.exists("database"):
             shutil.rmtree("database")
         if not isinstance(config.MONGO_URI, str):
@@ -70,4 +65,4 @@ class Telegram(Client):
             raise ValueError("No STRING session provided\n\nAdd STRING session in .env")
 
 
-client = Telegram()
+client: Telegram = Telegram()
